+++ src/xtd.forms.native.cocoa/src/xtd/forms/native/cocoa/file_dialog.mm
@@ -153,13 +153,13 @@ namespace {
     std::vector<std::string> filter_patterns = strings::split(filter, {'|'});
     std::vector<std::pair<std::string, std::vector<std::string>>> filters;
     if (filter_patterns.size() % 2 != 0) throw std::invalid_argument("Filter bad format");
-    for (int32_t index = 0; index < filter_patterns.size(); index += 2)
+    for (int_least32_t index = 0; index < filter_patterns.size(); index += 2)
       filters.push_back({filter_patterns[index], strings::split(filter_patterns[index + 1], {';'})});
     return filters;
   }
 }
 
-bool file_dialog::run_open_dialog(intptr_t hwnd, const std::string& default_ext, std::string& file_name, std::vector<std::string>& file_names, const std::string& filter, size_t filter_index, const std::string& initial_directory, size_t options, bool support_multi_dotted_extensions, const std::string& title) {
+bool file_dialog::run_open_dialog(intmax_t hwnd, const std::string& default_ext, std::string& file_name, std::vector<std::string>& file_names, const std::string& filter, size_t filter_index, const std::string& initial_directory, size_t options, bool support_multi_dotted_extensions, const std::string& title) {
   NSOpenPanel* openPanel = [[[NSOpenPanel alloc] init] autorelease];
   [openPanel setMessage:[NSString stringWithUTF8String:title.c_str()]];
   [openPanel setCanChooseFiles:YES];
@@ -181,8 +181,8 @@ bool file_dialog::run_open_dialog(intptr_t hwnd, const std::string& default_ext,
     file_name = [[(NSURL*)[[openPanel URLs] objectAtIndex:0] path] UTF8String];
   else {
     NSArray* urls = [openPanel URLs];
-    std::vector<std::string> fileNames((int32_t)[urls count]);
-    for (int32_t index = 0; index < (int32_t)[urls count]; index++)
+    std::vector<std::string> fileNames((int_least32_t)[urls count]);
+    for (int_least32_t index = 0; index < (int_least32_t)[urls count]; index++)
       fileNames[index] = [[(NSURL*)[urls objectAtIndex:index] path] UTF8String];
     file_name = fileNames[0];
     file_names = fileNames;
@@ -190,7 +190,7 @@ bool file_dialog::run_open_dialog(intptr_t hwnd, const std::string& default_ext,
   return true;
 }
 
-bool file_dialog::run_save_dialog(intptr_t hwnd, const std::string& default_ext, std::string& file_name, std::vector<std::string>& file_names, const std::string& filter, size_t filter_index, const std::string& initial_directory, size_t options, bool support_multi_dotted_extensions, const std::string& title) {
+bool file_dialog::run_save_dialog(intmax_t hwnd, const std::string& default_ext, std::string& file_name, std::vector<std::string>& file_names, const std::string& filter, size_t filter_index, const std::string& initial_directory, size_t options, bool support_multi_dotted_extensions, const std::string& title) {
   NSSavePanel* savePanel = [[[NSSavePanel alloc] init] autorelease];
   [savePanel setMessage:[NSString stringWithUTF8String:title.c_str()]];
   [savePanel setCanCreateDirectories:YES];
